- move DonationReminderHelper.kt to donationreminder package - lint fixes
